+++ src/ServiceProvider.php
@@ -13,7 +13,6 @@ use Laravel\Passkeys\Events\PasskeyVerified;
 use Laravel\Passkeys\Passkeys;
 use Livewire\Livewire;
 use NickDeKruijk\Leap\Commands\ModuleCommand;
-use NickDeKruijk\Leap\Commands\TemplateCommand;
 use NickDeKruijk\Leap\Commands\UserCommand;
 use NickDeKruijk\Leap\Middleware\Auth2FA;
 use NickDeKruijk\Leap\Middleware\LeapAuth;
@@ -119,7 +118,6 @@ class ServiceProvider extends \Illuminate\Support\ServiceProvider
         if ($this->app->runningInConsole()) {
             $this->commands([
                 ModuleCommand::class,
-                TemplateCommand::class,
                 UserCommand::class,
             ]);
         }